+++ tests/Driver/DatabaseDriverTest.php
@@ -618,9 +618,6 @@ class DatabaseDriverTest extends MockeryTestCase
         $driver->rollBackTransaction(null);
     }
 
-    /**
-     * @doesNotPerformAssertions
-     */
     public function testCommitTransactionExistingTransaction()
     {
         $db = Mockery::mock(QueryBuilder::class);